test(api): add SessionRepository specs
